+++ test/test_scenario4.py
@@ -18,8 +18,12 @@ def test_yaml_it_src_to_out_4():
 
         for file in output_root.rglob("*.yml"):
             output = file.read_text(encoding="utf-8")
-            assert ".sh" not in output or ". before_script.sh" in output
+            for line in output.split("\n"):
+                if ">>>" not in line and "<<<" not in line:
+                    assert ".sh" not in line or ". before_script.sh" in line
 
         for file in output_root.rglob("*.yaml"):
             output = file.read_text(encoding="utf-8")
-            assert ".sh" not in output or ". before_script.sh" in output
+            for line in output.split("\n"):
+                if ">>>" not in line and "<<<" not in line:
+                    assert ".sh" not in line or ". before_script.sh" in line