+++ test/test_scenario7.py
@@ -19,20 +19,26 @@ def test_yaml_it_src_to_out_7():
         found = 0
         for file in output_root.rglob("*.yml"):
             output = file.read_text(encoding="utf-8")
-            assert ".sh" not in output or ". before_script.sh" in output
+            for line in output.split("\n"):
+                if ">>>" not in line and "<<<" not in line:
+                    assert ".sh" not in line or ". before_script.sh" in line
             found += 1
         assert found
 
         found = 0
         for file in output_root.glob("templates/*.yml"):
             output = file.read_text(encoding="utf-8")
-            assert ".sh" not in output or ". before_script.sh" in output
+            for line in output.split("\n"):
+                if ">>>" not in line and "<<<" not in line:
+                    assert ".sh" not in line or ". before_script.sh" in line
             found += 1
         assert found
 
         found = 0
         for file in output_root.glob("templates/sub_template/*.yml"):
             output = file.read_text(encoding="utf-8")
-            assert ".sh" not in output or ". before_script.sh" in output
+            for line in output.split("\n"):
+                if ">>>" not in line and "<<<" not in line:
+                    assert ".sh" not in line or ". before_script.sh" in line
             found += 1
         assert found